has been add background.js as module;
