+++ static/manifest.json
@@ -6,8 +6,6 @@
   "options_page": "options.html",
   "background": {
     "scripts": [
-      "common.js",
-      "news.js",
       "background.js"
     ]
   },
