+++ static/news.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Українські новини</title>
-    <script src="../src/common.js"></script>
+    <script src="../src/js/common.js"></script>
     <script src="../src/news.js"></script>
     <script src="../src/tracker.js"></script>
     <link type="text/css" rel="stylesheet" href="style.css" media="all">
