+++ static/options.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Українські новини - налаштування</title>
-    <script src="../src/common.js"></script>
+    <script src="../src/js/common.js"></script>
     <link type="text/css" rel="stylesheet" href="options.css" media="all">
 </head>
 <body>
